[clang] Do not infer nounwind for const/pure/noalias.
The const and pure attributes are GCC extensions and should generally be
kept compatible with GCC. GCC permits them to throw exceptions: although
calls may be optimized away if the result is not used or not needed, if
they do get called and if they do throw exceptions, those exceptions are
handled as normal.

The noalias attribute is a MSVC extension and likewise, MSVC does not
optimize on the assumption that noalias functions do not throw.

To ensure that other optimizations, particularly removal of calls where
the result is unused, a new function attribute NoSideEffects is added.

Fixes #50365.