[BOLT] Gadget scanner: check function clobber list
Function outlining may result in calls to functions where in order to
provide accurate results, it is important that we do not flag them as
possibly clobbering registers that were never written to.

This requires updates to tests that were calling empty functions and
expecting that to be processed as clobbering all registers. These tests
are updated to perform an unknown syscall instead.