[BOLT] Take unmodelled side effects into account.
If we encounter an instruction with unmodelled side effects, we do not
know what registers will be used or changed. In particular, if we
execute a syscall, we know that arbitrary registers may be used or
changed depending on the exact syscall, and we need to conservatively
assume registers are no longer trusted afterwards.